fix(db): parametros int64 no se truncan en db_substitute_params ni al leer filas MySQL; tests de inyeccion SQL
Fase 4 (seguridad/datos de la capa SQL):
- db_params.c: 'vi' era int y se formateaba con %d -> un @param de 3000000000 llegaba a
  MySQL/Postgres como -1294967296 (afectaba IDs/timestamps/centavos grandes en el ERP).
  Ahora long long + %lld; se quitan los (int) al materializar params y en atoll.
- mysql_bridge.c: filas numericas se leian con atoi -> atoll.
- tests/db/test_db_params.c: casos injection_string (cierre de comilla + DROP), injection_or_1eq1
  (OR 1=1 llega como string -> siempre citado) e int64_param. 11/11.
- tests/lang/12_db/sqlite_params_injection.te: 4 vectores (DROP, OR 1=1, comentario --, UNION)
  neutralizados por el binding nativo del plugin SQLite.
Veredicto revisado de la auditoria: la capa de parametros (db_substitute_params) es
'emulated prepared statements' (escape del driver + comillas, numeros tipados, NULL) y no
interpola valores crudos: el riesgo de inyeccion via @param es bajo; server-side prepared
statements quedan como mejora, no como bloqueante.
Regresion Win: lang 99/99 (+3 xfail), api 9/9, tests/db 11/11.

+++ src/db_params.c
@@ -172,7 +172,7 @@ static ASTNode* db_value_to_typed_leaf(ASTNode* val) {
         Variable* v = find_variable(val->id);
         if (!v) return create_ast_leaf("NULL", 0, NULL, NULL);
         if (v->vtype == VAL_INT)
-            return create_ast_leaf_number("INT", (int)v->value.int_value, NULL, NULL);
+            return create_ast_leaf_number("INT", v->value.int_value, NULL, NULL);
         if (v->vtype == VAL_FLOAT) {
             char b[64]; te_fmt_double(b, sizeof(b), v->value.float_value);
             return create_ast_leaf("DB_RAW", 0, b, NULL);
@@ -192,7 +192,7 @@ static ASTNode* db_value_to_typed_leaf(ASTNode* val) {
         Variable* r = find_variable("__ret__");
         ASTNode* leaf;
         if (r && r->vtype == VAL_INT)
-            leaf = create_ast_leaf_number("INT", (int)r->value.int_value, NULL, NULL);
+            leaf = create_ast_leaf_number("INT", r->value.int_value, NULL, NULL);
         else if (r && r->vtype == VAL_FLOAT) {
             char b[64]; te_fmt_double(b, sizeof(b), r->value.float_value);
             leaf = create_ast_leaf("DB_RAW", 0, b, NULL);
@@ -219,7 +219,7 @@ static ASTNode* db_value_to_typed_leaf(ASTNode* val) {
             if (strpbrk(s, ".eE"))
                 leaf = create_ast_leaf("DB_RAW", 0, s, NULL);
             else
-                leaf = create_ast_leaf_number("INT", (int)atoll(s), NULL, NULL);
+                leaf = create_ast_leaf_number("INT", atoll(s), NULL, NULL);
         } else {
             leaf = create_ast_leaf("STRING", 0, s, NULL);
         }
@@ -291,7 +291,7 @@ static int append_value(char** buf, size_t* len, size_t* cap,
     }
     /* Resolver IDENTIFIER → variable */
     const char* tipo = val->type;
-    int     vi = 0;
+    long long vi = 0;
     double  vf = 0;
     const char* vs = NULL;
     int kind = 0; /* 1=int 2=float 3=str 4=null */
@@ -410,7 +410,7 @@ static int append_value(char** buf, size_t* len, size_t* cap,
     int n;
     switch (kind) {
         case 1:
-            n = snprintf(tmp, sizeof(tmp), "%d", vi);
+            n = snprintf(tmp, sizeof(tmp), "%lld", vi);
             buf_append(buf, len, cap, tmp, (size_t)n);
             return 1;
         case 2:
